some data table additions and cards to table change
