+++ index.html
@@ -18,7 +18,7 @@
 </body>
 <div></div>
 
-<div style="padding-top:10px; padding-bottom:10%"><strong>Hive of Heaven</strong> - Community Powered Night Sky Finder
+<div style="padding-top:10px; padding-bottom:10%"><strong>Hive of Heaven</strong> - Signs from Heaven
 </div>
