+++ src/App.css
@@ -808,7 +808,16 @@ p {
 .dark .indexButtonsPRESERVECOMMUNITYOBSERVATORY {
   background-color: #000000;
   color: red;
-  border: solid red
+  border: solid red;
+  margin-bottom: 3px;
+}
+
+.light .indexButtonsPRESERVECOMMUNITYOBSERVATORY {
+  background-color: #ececec;
+  color: rgb(0, 0, 0);
+  border: solid rgb(0, 174, 255);
+  margin-bottom: 3px;
+
 }
 
 .light .registerbtn {
@@ -862,4 +871,25 @@ p {
 
 input {
   border-radius: 10px;
+}
+
+
+
+.cardIndex{
+  width: 100%;
+  height: 150px;
+  background-color: #0483a3c0;
+
+}
+.cardIndex :hover{
+  width: 100%;
+  height: 150px;
+  background-color: #ffffff;
+
+}
+.imageIndex{
+  width: 150px;
+  height:"auto";
+border-radius: 10px;
+border: solid black;
 }
