+++ src/pages/TermsConditions.jsx
@@ -4,11 +4,16 @@ function TermsConditions(){
 return(
 
 <div>
+    <br></br>
+    <h1>Terms and Conditions</h1>
+    <br></br>
+    <br></br>
+    <div style={{margin:"30px"}}>
     We are not responsible for accuracy of posts. Posts are community driven and can only be authenticated by other users on the platform via upVoting, reviews, pictures, geolocation capturing, etc.
 
-    Any information on this site is not private, as it is stored on external servers. 
+    Any information on this site is not private. 
 
-    Thank you for understanding.
+    Thank you for understanding.</div>
 </div>
 
 )
